Use kernel interface info from drm/xgi_drm.h instead of xgi_driver.h.

+++ src/xgi_misc.h
@@ -42,8 +42,7 @@ extern Bool XGIPcieMemAllocate(ScrnInfoPtr pScrn, size_t size,
 extern Bool XGIPcieMemFree(ScrnInfoPtr pScrn, size_t size,
     unsigned long bufBusAddr, void *pBufVirtAddr);
 
-extern Bool  XGIShareAreaInfo(ScrnInfoPtr pScrn,
-                             unsigned long busAddr,
-                             unsigned long size);
+extern Bool  XGIShareAreaInfo(ScrnInfoPtr pScrn, unsigned long busAddr,
+    unsigned int size);
 
 #endif
